requirements.txt
requests
python-dotenv
typer[all]
playwright

